Change the style of the sidebar
I found a new website from which to download svg icon files.
I can edit the vectors in the website and that's very helpful.

I also discovered that I can use 'auto' to define a size for my rows
and columns in a grid. Basically, this will make the row/column as big
as its content.